[refactor] コンポーネントの分離とメモ化
