+++ src/pages/Top.tsx
@@ -1,44 +1,20 @@
-import { KeyboardEvent, MouseEvent, useState } from 'react';
 import { useDispatch } from 'react-redux';
 import { useHistory, useParams } from 'react-router-dom';
 import { useTranslation } from 'react-i18next';
 import { toast } from 'react-toastify';
-import {
-  AppBar,
-  Box,
-  Container,
-  Divider,
-  Drawer,
-  IconButton,
-  List,
-  ListItemIcon,
-  ListItemText,
-  makeStyles,
-  Toolbar,
-  Typography,
-} from '@material-ui/core';
-import { AccountBox, GitHub, Menu } from '@material-ui/icons';
-import clsx from 'clsx';
+import { AppBar, Box, Container, makeStyles, Toolbar } from '@material-ui/core';
 import 'date-fns';
 import * as util from 'util';
 
 import './Top.css';
-import {
-  SYSTEM_MAX_DATE,
-  SYSTEM_MIN_DATE,
-  LOG_E_INVALID_FORMAT,
-  LOG_E_OUT_OF_RANGE,
-  URL_CREATORPAGE,
-  URL_REPOSITORY,
-} from '../constant';
+import { SYSTEM_MAX_DATE, SYSTEM_MIN_DATE, LOG_E_INVALID_FORMAT, LOG_E_OUT_OF_RANGE } from '../constant';
 import { setDate } from '../stores/slices/viewSlice';
 import { formatDate } from '../utils/date';
-import ListItemLink from '../components/atoms/ListItemLink';
+import AppLogo from '../components/AppLogo';
 import Copyright from '../components/Copyright';
 import DateInput from '../components/DateInput';
 import DateResult from '../components/DateResult';
-
-type Anchor = 'top' | 'left' | 'bottom' | 'right';
+import DrawerMenu from '../components/DrawerMenu';
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -47,29 +23,12 @@ const useStyles = makeStyles((theme) => ({
   header: {
     flexGrow: 1,
   },
-  menuButton: {
-    marginRight: theme.spacing(2),
-  },
-  menuList: {
-    width: 220,
-  },
-  menuListFull: {
-    width: 'auto',
-  },
-  menuTopBox: {
-    display: 'flex',
-    margin: theme.spacing(3),
-  },
-  titleBox: {
+  title: {
     flexGrow: 1,
     display: 'none',
     [theme.breakpoints.up('sm')]: {
       display: 'flex',
     },
-  },
-  appIcon: {
-    width: 32,
-    height: 32,
   },
   content: {
     flexGrow: 1,
@@ -88,9 +47,6 @@ const Top = () => {
   const history = useHistory();
   const classes = useStyles();
   const [t] = useTranslation();
-
-  const anchor: Anchor = 'left';
-  const [anchorState, setAnchorState] = useState({ top: false, left: false, bottom: false, right: false });
 
   const params = useParams<{ date?: string }>();
   console.info(`Info: params.date = ${params.date}`);
@@ -111,61 +67,12 @@ const Top = () => {
   console.info(`Info: TargetDate = ${targetDate}`);
   dispatch(setDate(targetDate));
 
-  const toggleDrawer = (anchor: Anchor, open: boolean) => (event: KeyboardEvent | MouseEvent) => {
-    var e = event as KeyboardEvent;
-    if (e.type === 'keydown' && (e.key === 'Tab' || e.key === 'Shift')) return;
-    setAnchorState({ ...anchorState, [anchor]: open });
-  };
-
-  const appLogo = (className: string) => (
-    <Box className={className}>
-      <img className={classes.appIcon} src={`${process.env.PUBLIC_URL}/logo192.png`} alt="logo" loading="lazy" />
-      <div style={{ width: 20 }} />
-      <Typography variant="h6" noWrap>
-        {t('label.app-title')}
-      </Typography>
-    </Box>
-  );
-
-  const menuList = (anchor: Anchor) => (
-    <Box
-      className={clsx(classes.menuList, {
-        [classes.menuListFull]: anchor === 'top' || anchor === 'bottom',
-      })}
-      role="presentation"
-      onClick={toggleDrawer(anchor, false)}
-      onKeyDown={toggleDrawer(anchor, false)}
-    >
-      {appLogo(classes.menuTopBox)}
-      <Divider />
-      <List>
-        <ListItemLink href={URL_REPOSITORY} target="_blank">
-          <ListItemIcon>
-            <GitHub />
-          </ListItemIcon>
-          <ListItemText primary={t('label.repository')} />
-        </ListItemLink>
-        <ListItemLink href={URL_CREATORPAGE} target="_blank">
-          <ListItemIcon>
-            <AccountBox />
-          </ListItemIcon>
-          <ListItemText primary={t('label.creator')} />
-        </ListItemLink>
-      </List>
-    </Box>
-  );
-
   return (
     <Box className={classes.root}>
       <AppBar className={classes.header} position="sticky" color="default">
         <Toolbar>
-          <IconButton className={classes.menuButton} edge="start" onClick={toggleDrawer(anchor, true)}>
-            <Menu />
-          </IconButton>
-          <Drawer anchor={anchor} open={anchorState[anchor]} onClose={toggleDrawer(anchor, false)}>
-            {menuList(anchor)}
-          </Drawer>
-          {appLogo(classes.titleBox)}
+          <DrawerMenu anchor="left" />
+          <AppLogo className={classes.title} />
           <div style={{ flex: '1 0 0' }} />
           <DateInput minDate={SYSTEM_MIN_DATE} maxDate={SYSTEM_MAX_DATE} />
         </Toolbar>
